git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@31441 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ core_modules/cache/lib/cacheLib.class.php
@@ -79,9 +79,33 @@ class cacheLib
     /**
      * Delete all cached file's of the cache system   
      */
-    function _deleteAllFiles()
+    function _deleteAllFiles($cacheEngine = null)
     {
-        \Env::get('cache')->deleteAll();
+        if (!in_array($cacheEngine, array('cxPages', 'cxEntries'))) {
+            \Env::get('cache')->deleteAll();
+            return;
+        }
+        $handleDir = opendir($this->strCachePath);
+        if ($handleDir) {
+            while ($strFile = readdir($handleDir)) {
+                if ($strFile != '.' && $strFile != '..') {
+                    switch ($cacheEngine) {
+                        case 'cxPages':
+                            if(is_file($this->strCachePath . $strFile)){
+                                unlink($this->strCachePath . $strFile);
+                            }
+                            break;
+                        case 'cxEntries':
+                            \Env::get('cache')->deleteAll();
+                            break;
+                        default:
+                            unlink($this->strCachePath . $strFile);
+                            break;
+                    }
+                }
+            }
+            closedir($handleDir);
+        }
     }
     
     protected function initOPCaching() {
